Simplify json editor
